[WIP] start re-organizing variables into task module, splitting API into group/participant
