+++ AFQ/tasks/mapping.py
@@ -9,6 +9,10 @@ from AFQ.tasks.decorators import as_file
 from AFQ.tasks.utils import get_fname, with_name
 import AFQ.data as afd
 import AFQ.utils.volume as auv
+from AFQ.definitions.mapping import SynMap
+from AFQ.definitions.utils import Definition
+from AFQ.definitions.mask import FullMask
+import AFQ.api.bundle_dict as abd
 
 from dipy.io.streamline import load_tractogram
 from dipy.io.stateful_tractogram import Space
@@ -88,14 +92,23 @@ def export_rois(subses_dict, bundle_dict, mapping, dwi_affine):
 
 
 @pimms.calc("mapping")
-def mapping(subses_dict, mapping_definition, reg_subject, reg_template):
+def mapping(subses_dict, reg_subject, reg_template,
+            mapping_definition=SynMap()):
+    if not isinstance(mapping, Definition):
+        raise TypeError(
+            "mapping must be a mapping defined"
+            + " in `AFQ.definitions.mapping`")
     return mapping_definition.get_for_subses(
         subses_dict, reg_subject, reg_template)
 
 
 @pimms.calc("mapping")
-def sls_mapping(subses_dict, mapping_definition, reg_subject, reg_template,
-                tractography_imap):
+def sls_mapping(subses_dict, reg_subject, reg_template,
+                tractography_imap, mapping_definition=SynMap()):
+    if not isinstance(mapping, Definition):
+        raise TypeError(
+            "mapping must be a mapping defined"
+            + " in `AFQ.definitions.mapping`")
     streamlines_file = tractography_imap["streamlines_file"]
     tg = load_tractogram(
         streamlines_file, reg_subject,
@@ -137,13 +150,84 @@ def get_reg_subject(reg_subject_spec, data_imap):
     return img
 
 
-def get_mapping_plan(reg_subject, scalars, use_sls=False):
+def get_mapping_plan(kwargs, use_sls=False):
+    if "bundle_info" in kwargs and not ((
+            isinstance(kwargs["bundle_info"], list)
+            and isinstance(kwargs["bundle_info"][0], str)) or (
+                isinstance(kwargs["bundle_info"], dict)) or (
+                    isinstance(kwargs["bundle_info"], abd.BundleDict))):
+        raise TypeError((
+            "bundle_info must be a list of strings,"
+            " a dict, or a BundleDict"))
+    if "reg_template" in kwargs and\
+        not isinstance(kwargs["reg_template"], str)\
+            and not isinstance(kwargs["reg_template"], nib.Nifti1Image):
+        raise TypeError(
+            "reg_template must be a str or Nifti1Image")
+    if "reg_template" in kwargs and\
+        not isinstance(kwargs["reg_subject"], str)\
+            and not isinstance(kwargs["reg_subject"], nib.Nifti1Image)\
+            and not isinstance(kwargs["reg_subject"], dict):
+        raise TypeError(
+            "reg_subject must be a str, dict, or Nifti1Image")
+
     mapping_tasks = with_name([
         export_registered_b0, template_xform, export_rois, mapping,
         get_reg_subject])
 
+    if "reg_template" not in kwargs:
+        kwargs["reg_template"] = "mni_T1"
+    if "reg_subject" not in kwargs:
+        kwargs["reg_subject"] = "power_map"
+
+    if isinstance(kwargs["reg_subject"], str) and\
+        isinstance(kwargs["reg_template"], str)\
+            and (kwargs["reg_subject"].lower() == 'subject_sls'
+                 or kwargs["reg_template"].lower() == 'hcp_atlas'):
+        if kwargs["reg_template"].lower() != 'hcp_atlas':
+            raise TypeError(
+                "If reg_subject is 'subject_sls',"
+                + " reg_template must be 'hcp_atlas'")
+        if kwargs["reg_subject"].lower() != 'subject_sls':
+            raise TypeError(
+                "If reg_template is 'hcp_atlas',"
+                + " reg_subject must be 'subject_sls'")
+
+    seg_algo = kwargs.get(kwargs["segmentation_params"], None)
+    if seg_algo is not None:
+        seg_algo = kwargs.get(kwargs["seg_algo"], None)
+
+    if "bundle_info" in kwargs and not ((
+            isinstance(kwargs["bundle_info"], list)
+            and isinstance(kwargs["bundle_info"][0], str)) or (
+                isinstance(kwargs["bundle_info"], dict)) or (
+                    isinstance(kwargs["bundle_info"], abd.BundleDict))):
+        raise TypeError((
+            "bundle_info must a list of strings,"
+            " a dict, or a BundleDict"))
+
+    if "bundle_info" not in kwargs:
+        if seg_algo == "reco" or seg_algo == "reco16":
+            kwargs["bundle_info"] = abd.RECO_BUNDLES_16
+        elif seg_algo == "reco80":
+            kwargs["bundle_info"] = abd.RECO_BUNDLES_80
+        else:
+            kwargs["bundle_info"] = abd.BUNDLES
+
+    # set kwargs["reg_template"] and kwargs["bundle_info"]:
+    kwargs["reg_template"] = get_reg_template(
+        kwargs["reg_template"],
+        ~isinstance(kwargs["brain_mask_definition"], FullMask))
+    if isinstance(kwargs["bundle_info"], abd.BundleDict):
+        kwargs["bundle_dict"] = kwargs["bundle_info"]
+    else:
+        kwargs["bundle_dict"] = abd.BundleDict(
+            kwargs["bundle_info"],
+            seg_algo=seg_algo,
+            resample_to=kwargs["reg_template"])
+
     # add custom scalars
-    for scalar in scalars:
+    for scalar in kwargs["scalars"]:
         if not isinstance(scalar, str):
             mapping_tasks["{scalar.name}_file_res"] =\
                 pimms.calc(f"{scalar.name}_file")(scalar.get_for_subses())
@@ -152,3 +236,24 @@ def get_mapping_plan(reg_subject, scalars, use_sls=False):
         mapping_tasks["mapping_res"] = sls_mapping
 
     return pimms.plan(**mapping_tasks)
+
+
+def get_reg_template(reg_template, mask_template):
+    if isinstance(reg_template, nib.Nifti1Image):
+        return reg_template
+
+    img_l = reg_template.lower()
+    if img_l == "mni_t2":
+        img = afd.read_mni_template(
+            mask=mask_template, weight="T2w")
+    elif img_l == "mni_t1":
+        img = afd.read_mni_template(
+            mask=mask_template, weight="T1w")
+    elif img_l == "dti_fa_template":
+        img = afd.read_ukbb_fa_template(mask=mask_template)
+    elif img_l == "hcp_atlas":
+        img = afd.read_mni_template(mask=mask_template)
+    else:
+        img = nib.load(reg_template)
+
+    return img
